divide JobProxy in multiple modules #1
