+++ JobProxyModel/src/main/java/de/unibi/cebitec/bibiserv/jobproxy/model/state/ObjectFactory.java
@@ -1,0 +1,54 @@
+package de.unibi.cebitec.bibiserv.jobproxy.model.state;//
+// This file was generated by the JavaTM Architecture for XML Binding(JAXB) Reference Implementation, v2.2.8-b130911.1802 
+// See <a href="http://java.sun.com/xml/jaxb">http://java.sun.com/xml/jaxb</a> 
+// Any modifications to this file will be lost upon recompilation of the source schema. 
+// Generated on: 2016.03.01 at 12:26:27 PM CET 
+//
+
+
+
+import javax.xml.bind.annotation.XmlRegistry;
+
+
+/**
+ * This object contains factory methods for each 
+ * Java content interface and Java element interface 
+ * generated in the de.unibi.cebitec.bibiserv.jobproxy.data.state package. 
+ * <p>An ObjectFactory allows you to programatically 
+ * construct new instances of the Java representation 
+ * for XML content. The Java representation of XML 
+ * content can consist of schema derived interfaces 
+ * and classes representing the binding of schema 
+ * type definitions, element declarations and model 
+ * groups.  Factory methods for each of these are 
+ * provided in this class.
+ * 
+ */
+@XmlRegistry
+public class ObjectFactory {
+
+
+    /**
+     * Create a new ObjectFactory that can be used to create new instances of schema derived classes for package: de.unibi.cebitec.bibiserv.jobproxy.data.state
+     * 
+     */
+    public ObjectFactory() {
+    }
+
+    /**
+     * Create an instance of {@link State }
+     * 
+     */
+    public State createState() {
+        return new State();
+    }
+
+    /**
+     * Create an instance of {@link States }
+     * 
+     */
+    public States createStates() {
+        return new States();
+    }
+
+}
