+++ JobProxyModel/src/main/java/de/unibi/cebitec/bibiserv/jobproxy/model/task/TMounts.java
@@ -1,0 +1,163 @@
+package de.unibi.cebitec.bibiserv.jobproxy.model.task;//
+// This file was generated by the JavaTM Architecture for XML Binding(JAXB) Reference Implementation, v2.2.8-b130911.1802 
+// See <a href="http://java.sun.com/xml/jaxb">http://java.sun.com/xml/jaxb</a> 
+// Any modifications to this file will be lost upon recompilation of the source schema. 
+// Generated on: 2016.03.01 at 12:26:26 PM CET 
+//
+
+
+
+import javax.xml.bind.annotation.XmlAccessType;
+import javax.xml.bind.annotation.XmlAccessorType;
+import javax.xml.bind.annotation.XmlElement;
+import javax.xml.bind.annotation.XmlType;
+
+
+/**
+ * Describe a list of host,container mount values
+ * 
+ * <p>Java class for tMounts complex type.
+ * 
+ * <p>The following schema fragment specifies the expected content contained within this class.
+ * 
+ * <pre>
+ * &lt;complexType name="tMounts">
+ *   &lt;complexContent>
+ *     &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType">
+ *       &lt;sequence>
+ *         &lt;element name="mount">
+ *           &lt;complexType>
+ *             &lt;complexContent>
+ *               &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType">
+ *                 &lt;all>
+ *                   &lt;element name="host" type="{http://www.w3.org/2001/XMLSchema}string"/>
+ *                   &lt;element name="container" type="{http://www.w3.org/2001/XMLSchema}string"/>
+ *                 &lt;/all>
+ *               &lt;/restriction>
+ *             &lt;/complexContent>
+ *           &lt;/complexType>
+ *         &lt;/element>
+ *       &lt;/sequence>
+ *     &lt;/restriction>
+ *   &lt;/complexContent>
+ * &lt;/complexType>
+ * </pre>
+ * 
+ * 
+ */
+@XmlAccessorType(XmlAccessType.FIELD)
+@XmlType(name = "tMounts", propOrder = {
+    "mount"
+})
+public class TMounts {
+
+    @XmlElement(required = true)
+    protected TMounts.Mount mount;
+
+    /**
+     * Gets the value of the mount property.
+     * 
+     * @return
+     *     possible object is
+     *     {@link TMounts.Mount }
+     *     
+     */
+    public TMounts.Mount getMount() {
+        return mount;
+    }
+
+    /**
+     * Sets the value of the mount property.
+     * 
+     * @param value
+     *     allowed object is
+     *     {@link TMounts.Mount }
+     *     
+     */
+    public void setMount(TMounts.Mount value) {
+        this.mount = value;
+    }
+
+
+    /**
+     * <p>Java class for anonymous complex type.
+     * 
+     * <p>The following schema fragment specifies the expected content contained within this class.
+     * 
+     * <pre>
+     * &lt;complexType>
+     *   &lt;complexContent>
+     *     &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType">
+     *       &lt;all>
+     *         &lt;element name="host" type="{http://www.w3.org/2001/XMLSchema}string"/>
+     *         &lt;element name="container" type="{http://www.w3.org/2001/XMLSchema}string"/>
+     *       &lt;/all>
+     *     &lt;/restriction>
+     *   &lt;/complexContent>
+     * &lt;/complexType>
+     * </pre>
+     * 
+     * 
+     */
+    @XmlAccessorType(XmlAccessType.FIELD)
+    @XmlType(name = "", propOrder = {
+
+    })
+    public static class Mount {
+
+        @XmlElement(required = true)
+        protected String host;
+        @XmlElement(required = true)
+        protected String container;
+
+        /**
+         * Gets the value of the host property.
+         * 
+         * @return
+         *     possible object is
+         *     {@link String }
+         *     
+         */
+        public String getHost() {
+            return host;
+        }
+
+        /**
+         * Sets the value of the host property.
+         * 
+         * @param value
+         *     allowed object is
+         *     {@link String }
+         *     
+         */
+        public void setHost(String value) {
+            this.host = value;
+        }
+
+        /**
+         * Gets the value of the container property.
+         * 
+         * @return
+         *     possible object is
+         *     {@link String }
+         *     
+         */
+        public String getContainer() {
+            return container;
+        }
+
+        /**
+         * Sets the value of the container property.
+         * 
+         * @param value
+         *     allowed object is
+         *     {@link String }
+         *     
+         */
+        public void setContainer(String value) {
+            this.container = value;
+        }
+
+    }
+
+}
