+++ JobProxyModel/src/main/java/de/unibi/cebitec/bibiserv/jobproxy/model/task/TPorts.java
@@ -1,0 +1,145 @@
+package de.unibi.cebitec.bibiserv.jobproxy.model.task;//
+// This file was generated by the JavaTM Architecture for XML Binding(JAXB) Reference Implementation, v2.2.8-b130911.1802 
+// See <a href="http://java.sun.com/xml/jaxb">http://java.sun.com/xml/jaxb</a> 
+// Any modifications to this file will be lost upon recompilation of the source schema. 
+// Generated on: 2016.03.01 at 12:26:26 PM CET 
+//
+
+
+
+import javax.xml.bind.annotation.XmlAccessType;
+import javax.xml.bind.annotation.XmlAccessorType;
+import javax.xml.bind.annotation.XmlElement;
+import javax.xml.bind.annotation.XmlType;
+
+
+/**
+ * Describe a list of host,container port values
+ * 
+ * <p>Java class for tPorts complex type.
+ * 
+ * <p>The following schema fragment specifies the expected content contained within this class.
+ * 
+ * <pre>
+ * &lt;complexType name="tPorts">
+ *   &lt;complexContent>
+ *     &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType">
+ *       &lt;sequence>
+ *         &lt;element name="port">
+ *           &lt;complexType>
+ *             &lt;complexContent>
+ *               &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType">
+ *                 &lt;all>
+ *                   &lt;element name="host" type="{http://www.w3.org/2001/XMLSchema}int"/>
+ *                   &lt;element name="container" type="{http://www.w3.org/2001/XMLSchema}int"/>
+ *                 &lt;/all>
+ *               &lt;/restriction>
+ *             &lt;/complexContent>
+ *           &lt;/complexType>
+ *         &lt;/element>
+ *       &lt;/sequence>
+ *     &lt;/restriction>
+ *   &lt;/complexContent>
+ * &lt;/complexType>
+ * </pre>
+ * 
+ * 
+ */
+@XmlAccessorType(XmlAccessType.FIELD)
+@XmlType(name = "tPorts", propOrder = {
+    "port"
+})
+public class TPorts {
+
+    @XmlElement(required = true)
+    protected TPorts.Port port;
+
+    /**
+     * Gets the value of the port property.
+     * 
+     * @return
+     *     possible object is
+     *     {@link TPorts.Port }
+     *     
+     */
+    public TPorts.Port getPort() {
+        return port;
+    }
+
+    /**
+     * Sets the value of the port property.
+     * 
+     * @param value
+     *     allowed object is
+     *     {@link TPorts.Port }
+     *     
+     */
+    public void setPort(TPorts.Port value) {
+        this.port = value;
+    }
+
+
+    /**
+     * <p>Java class for anonymous complex type.
+     * 
+     * <p>The following schema fragment specifies the expected content contained within this class.
+     * 
+     * <pre>
+     * &lt;complexType>
+     *   &lt;complexContent>
+     *     &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType">
+     *       &lt;all>
+     *         &lt;element name="host" type="{http://www.w3.org/2001/XMLSchema}int"/>
+     *         &lt;element name="container" type="{http://www.w3.org/2001/XMLSchema}int"/>
+     *       &lt;/all>
+     *     &lt;/restriction>
+     *   &lt;/complexContent>
+     * &lt;/complexType>
+     * </pre>
+     * 
+     * 
+     */
+    @XmlAccessorType(XmlAccessType.FIELD)
+    @XmlType(name = "", propOrder = {
+
+    })
+    public static class Port {
+
+        protected int host;
+        protected int container;
+
+        /**
+         * Gets the value of the host property.
+         * 
+         */
+        public int getHost() {
+            return host;
+        }
+
+        /**
+         * Sets the value of the host property.
+         * 
+         */
+        public void setHost(int value) {
+            this.host = value;
+        }
+
+        /**
+         * Gets the value of the container property.
+         * 
+         */
+        public int getContainer() {
+            return container;
+        }
+
+        /**
+         * Sets the value of the container property.
+         * 
+         */
+        public void setContainer(int value) {
+            this.container = value;
+        }
+
+    }
+
+}
