+++ JobProxyServer/pom.xml
@@ -1,0 +1,26 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <parent>
+        <groupId>de.unibi.cebitec.bibiserv.jkrueger</groupId>
+        <artifactId>JobProxy</artifactId>
+        <version>1.0-SNAPSHOT</version>
+    </parent>
+    <modelVersion>4.0.0</modelVersion>
+
+    <artifactId>JobProxyServer</artifactId>
+    <dependencies>
+        <dependency>
+            <groupId>de.unibi.cebitec.bibiserv.jkrueger</groupId>
+            <artifactId>JobProxyModel</artifactId>
+            <version>1.0-SNAPSHOT</version>
+        </dependency>
+        <dependency>
+            <groupId>de.unibi.cebitec.bibiserv.jkrueger</groupId>
+            <artifactId>JobProxyChronos</artifactId>
+            <version>1.0-SNAPSHOT</version>
+        </dependency>
+    </dependencies>
+
+</project>
