--- src/main/java/de/unibi/cebitec/bibiserv/jobproxy/data/chronos/ObjectFactory.java
@@ -1,95 +1,0 @@
-//
-// This file was generated by the JavaTM Architecture for XML Binding(JAXB) Reference Implementation, v2.2.8-b130911.1802 
-// See <a href="http://java.sun.com/xml/jaxb">http://java.sun.com/xml/jaxb</a> 
-// Any modifications to this file will be lost upon recompilation of the source schema. 
-// Generated on: 2016.03.01 at 12:26:28 PM CET 
-//
-
-
-package de.unibi.cebitec.bibiserv.jobproxy.data.chronos;
-
-import javax.xml.bind.annotation.XmlRegistry;
-
-
-/**
- * This object contains factory methods for each 
- * Java content interface and Java element interface 
- * generated in the de.unibi.cebitec.bibiserv.jobproxy.data.chronos package. 
- * <p>An ObjectFactory allows you to programatically 
- * construct new instances of the Java representation 
- * for XML content. The Java representation of XML 
- * content can consist of schema derived interfaces 
- * and classes representing the binding of schema 
- * type definitions, element declarations and model 
- * groups.  Factory methods for each of these are 
- * provided in this class.
- * 
- */
-@XmlRegistry
-public class ObjectFactory {
-
-
-    /**
-     * Create a new ObjectFactory that can be used to create new instances of schema derived classes for package: de.unibi.cebitec.bibiserv.jobproxy.data.chronos
-     * 
-     */
-    public ObjectFactory() {
-    }
-
-    /**
-     * Create an instance of {@link Jobconfig }
-     * 
-     */
-    public Jobconfig createJobconfig() {
-        return new Jobconfig();
-    }
-
-    /**
-     * Create an instance of {@link Jobconfig.EnvironmentVariables }
-     * 
-     */
-    public Jobconfig.EnvironmentVariables createJobconfigEnvironmentVariables() {
-        return new Jobconfig.EnvironmentVariables();
-    }
-
-    /**
-     * Create an instance of {@link Jobconfig.Container }
-     * 
-     */
-    public Jobconfig.Container createJobconfigContainer() {
-        return new Jobconfig.Container();
-    }
-
-    /**
-     * Create an instance of {@link Jobconfig.Constraints }
-     * 
-     */
-    public Jobconfig.Constraints createJobconfigConstraints() {
-        return new Jobconfig.Constraints();
-    }
-
-    /**
-     * Create an instance of {@link Tvolume }
-     * 
-     */
-    public Tvolume createTvolume() {
-        return new Tvolume();
-    }
-
-    /**
-     * Create an instance of {@link Jobconfig.EnvironmentVariables.Var }
-     * 
-     */
-    public Jobconfig.EnvironmentVariables.Var createJobconfigEnvironmentVariablesVar() {
-        return new Jobconfig.EnvironmentVariables.Var();
-    }
-
-    /**
-     * Create an instance of {@link Jobconfig.Container.Volumes }
-     * 
-     */
-    public Jobconfig.Container.Volumes createJobconfigContainerVolumes() {
-        return new Jobconfig.Container.Volumes();
-    }
-
-}
