--- src/main/java/de/unibi/cebitec/bibiserv/jobproxy/data/chronos/Tvolume.java
@@ -1,125 +1,0 @@
-//
-// This file was generated by the JavaTM Architecture for XML Binding(JAXB) Reference Implementation, v2.2.8-b130911.1802 
-// See <a href="http://java.sun.com/xml/jaxb">http://java.sun.com/xml/jaxb</a> 
-// Any modifications to this file will be lost upon recompilation of the source schema. 
-// Generated on: 2016.03.01 at 12:26:28 PM CET 
-//
-
-
-package de.unibi.cebitec.bibiserv.jobproxy.data.chronos;
-
-import javax.xml.bind.annotation.XmlAccessType;
-import javax.xml.bind.annotation.XmlAccessorType;
-import javax.xml.bind.annotation.XmlElement;
-import javax.xml.bind.annotation.XmlType;
-
-
-/**
- *  Specification of one Volume
- * 
- * <p>Java class for Tvolume complex type.
- * 
- * <p>The following schema fragment specifies the expected content contained within this class.
- * 
- * <pre>
- * &lt;complexType name="Tvolume">
- *   &lt;complexContent>
- *     &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType">
- *       &lt;all>
- *         &lt;element name="containerPath" type="{http://www.w3.org/2001/XMLSchema}string"/>
- *         &lt;element name="hostPath" type="{http://www.w3.org/2001/XMLSchema}string"/>
- *         &lt;element name="mode" type="{http://www.w3.org/2001/XMLSchema}string"/>
- *       &lt;/all>
- *     &lt;/restriction>
- *   &lt;/complexContent>
- * &lt;/complexType>
- * </pre>
- * 
- * 
- */
-@XmlAccessorType(XmlAccessType.FIELD)
-@XmlType(name = "Tvolume", propOrder = {
-
-})
-public class Tvolume {
-
-    @XmlElement(required = true)
-    protected String containerPath;
-    @XmlElement(required = true)
-    protected String hostPath;
-    @XmlElement(required = true)
-    protected String mode;
-
-    /**
-     * Gets the value of the containerPath property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link String }
-     *     
-     */
-    public String getContainerPath() {
-        return containerPath;
-    }
-
-    /**
-     * Sets the value of the containerPath property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link String }
-     *     
-     */
-    public void setContainerPath(String value) {
-        this.containerPath = value;
-    }
-
-    /**
-     * Gets the value of the hostPath property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link String }
-     *     
-     */
-    public String getHostPath() {
-        return hostPath;
-    }
-
-    /**
-     * Sets the value of the hostPath property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link String }
-     *     
-     */
-    public void setHostPath(String value) {
-        this.hostPath = value;
-    }
-
-    /**
-     * Gets the value of the mode property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link String }
-     *     
-     */
-    public String getMode() {
-        return mode;
-    }
-
-    /**
-     * Sets the value of the mode property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link String }
-     *     
-     */
-    public void setMode(String value) {
-        this.mode = value;
-    }
-
-}
