--- src/main/java/de/unibi/cebitec/bibiserv/jobproxy/data/task/ObjectFactory.java
@@ -1,87 +1,0 @@
-//
-// This file was generated by the JavaTM Architecture for XML Binding(JAXB) Reference Implementation, v2.2.8-b130911.1802 
-// See <a href="http://java.sun.com/xml/jaxb">http://java.sun.com/xml/jaxb</a> 
-// Any modifications to this file will be lost upon recompilation of the source schema. 
-// Generated on: 2016.03.01 at 12:26:26 PM CET 
-//
-
-
-package de.unibi.cebitec.bibiserv.jobproxy.data.task;
-
-import javax.xml.bind.annotation.XmlRegistry;
-
-
-/**
- * This object contains factory methods for each 
- * Java content interface and Java element interface 
- * generated in the de.unibi.cebitec.bibiserv.jobproxy.data.task package. 
- * <p>An ObjectFactory allows you to programatically 
- * construct new instances of the Java representation 
- * for XML content. The Java representation of XML 
- * content can consist of schema derived interfaces 
- * and classes representing the binding of schema 
- * type definitions, element declarations and model 
- * groups.  Factory methods for each of these are 
- * provided in this class.
- * 
- */
-@XmlRegistry
-public class ObjectFactory {
-
-
-    /**
-     * Create a new ObjectFactory that can be used to create new instances of schema derived classes for package: de.unibi.cebitec.bibiserv.jobproxy.data.task
-     * 
-     */
-    public ObjectFactory() {
-    }
-
-    /**
-     * Create an instance of {@link TMounts }
-     * 
-     */
-    public TMounts createTMounts() {
-        return new TMounts();
-    }
-
-    /**
-     * Create an instance of {@link TPorts }
-     * 
-     */
-    public TPorts createTPorts() {
-        return new TPorts();
-    }
-
-    /**
-     * Create an instance of {@link Task }
-     * 
-     */
-    public Task createTask() {
-        return new Task();
-    }
-
-    /**
-     * Create an instance of {@link TContainer }
-     * 
-     */
-    public TContainer createTContainer() {
-        return new TContainer();
-    }
-
-    /**
-     * Create an instance of {@link TMounts.Mount }
-     * 
-     */
-    public TMounts.Mount createTMountsMount() {
-        return new TMounts.Mount();
-    }
-
-    /**
-     * Create an instance of {@link TPorts.Port }
-     * 
-     */
-    public TPorts.Port createTPortsPort() {
-        return new TPorts.Port();
-    }
-
-}
