--- src/main/java/de/unibi/cebitec/bibiserv/jobproxy/data/task/TContainer.java
@@ -1,164 +1,0 @@
-//
-// This file was generated by the JavaTM Architecture for XML Binding(JAXB) Reference Implementation, v2.2.8-b130911.1802 
-// See <a href="http://java.sun.com/xml/jaxb">http://java.sun.com/xml/jaxb</a> 
-// Any modifications to this file will be lost upon recompilation of the source schema. 
-// Generated on: 2016.03.01 at 12:26:26 PM CET 
-//
-
-
-package de.unibi.cebitec.bibiserv.jobproxy.data.task;
-
-import java.util.ArrayList;
-import java.util.List;
-import javax.xml.bind.annotation.XmlAccessType;
-import javax.xml.bind.annotation.XmlAccessorType;
-import javax.xml.bind.annotation.XmlElement;
-import javax.xml.bind.annotation.XmlType;
-
-
-/**
- * Container description
- * 
- * <p>Java class for tContainer complex type.
- * 
- * <p>The following schema fragment specifies the expected content contained within this class.
- * 
- * <pre>
- * &lt;complexType name="tContainer">
- *   &lt;complexContent>
- *     &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType">
- *       &lt;sequence>
- *         &lt;element name="image" type="{http://www.w3.org/2001/XMLSchema}string"/>
- *         &lt;element name="ports" type="{}tPorts" maxOccurs="unbounded" minOccurs="0"/>
- *         &lt;element name="mounts" type="{}tMounts" maxOccurs="unbounded" minOccurs="0"/>
- *         &lt;element name="cmd" type="{http://www.w3.org/2001/XMLSchema}string" minOccurs="0"/>
- *       &lt;/sequence>
- *     &lt;/restriction>
- *   &lt;/complexContent>
- * &lt;/complexType>
- * </pre>
- * 
- * 
- */
-@XmlAccessorType(XmlAccessType.FIELD)
-@XmlType(name = "tContainer", propOrder = {
-    "image",
-    "ports",
-    "mounts",
-    "cmd"
-})
-public class TContainer {
-
-    @XmlElement(required = true)
-    protected String image;
-    protected List<TPorts> ports;
-    protected List<TMounts> mounts;
-    protected String cmd;
-
-    /**
-     * Gets the value of the image property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link String }
-     *     
-     */
-    public String getImage() {
-        return image;
-    }
-
-    /**
-     * Sets the value of the image property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link String }
-     *     
-     */
-    public void setImage(String value) {
-        this.image = value;
-    }
-
-    /**
-     * Gets the value of the ports property.
-     * 
-     * <p>
-     * This accessor method returns a reference to the live list,
-     * not a snapshot. Therefore any modification you make to the
-     * returned list will be present inside the JAXB object.
-     * This is why there is not a <CODE>set</CODE> method for the ports property.
-     * 
-     * <p>
-     * For example, to add a new item, do as follows:
-     * <pre>
-     *    getPorts().add(newItem);
-     * </pre>
-     * 
-     * 
-     * <p>
-     * Objects of the following type(s) are allowed in the list
-     * {@link TPorts }
-     * 
-     * 
-     */
-    public List<TPorts> getPorts() {
-        if (ports == null) {
-            ports = new ArrayList<TPorts>();
-        }
-        return this.ports;
-    }
-
-    /**
-     * Gets the value of the mounts property.
-     * 
-     * <p>
-     * This accessor method returns a reference to the live list,
-     * not a snapshot. Therefore any modification you make to the
-     * returned list will be present inside the JAXB object.
-     * This is why there is not a <CODE>set</CODE> method for the mounts property.
-     * 
-     * <p>
-     * For example, to add a new item, do as follows:
-     * <pre>
-     *    getMounts().add(newItem);
-     * </pre>
-     * 
-     * 
-     * <p>
-     * Objects of the following type(s) are allowed in the list
-     * {@link TMounts }
-     * 
-     * 
-     */
-    public List<TMounts> getMounts() {
-        if (mounts == null) {
-            mounts = new ArrayList<TMounts>();
-        }
-        return this.mounts;
-    }
-
-    /**
-     * Gets the value of the cmd property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link String }
-     *     
-     */
-    public String getCmd() {
-        return cmd;
-    }
-
-    /**
-     * Sets the value of the cmd property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link String }
-     *     
-     */
-    public void setCmd(String value) {
-        this.cmd = value;
-    }
-
-}
