--- src/main/java/de/unibi/cebitec/bibiserv/jobproxy/data/task/Task.java
@@ -1,262 +1,0 @@
-//
-// This file was generated by the JavaTM Architecture for XML Binding(JAXB) Reference Implementation, v2.2.8-b130911.1802 
-// See <a href="http://java.sun.com/xml/jaxb">http://java.sun.com/xml/jaxb</a> 
-// Any modifications to this file will be lost upon recompilation of the source schema. 
-// Generated on: 2016.03.01 at 12:26:26 PM CET 
-//
-
-
-package de.unibi.cebitec.bibiserv.jobproxy.data.task;
-
-import javax.xml.bind.annotation.XmlAccessType;
-import javax.xml.bind.annotation.XmlAccessorType;
-import javax.xml.bind.annotation.XmlElement;
-import javax.xml.bind.annotation.XmlRootElement;
-import javax.xml.bind.annotation.XmlType;
-
-
-/**
- * <p>Java class for anonymous complex type.
- * 
- * <p>The following schema fragment specifies the expected content contained within this class.
- * 
- * <pre>
- * &lt;complexType>
- *   &lt;complexContent>
- *     &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType">
- *       &lt;sequence>
- *         &lt;element name="user" type="{http://www.w3.org/2001/XMLSchema}string"/>
- *         &lt;element name="cores" type="{http://www.w3.org/2001/XMLSchema}int" minOccurs="0"/>
- *         &lt;element name="memory" type="{http://www.w3.org/2001/XMLSchema}int" minOccurs="0"/>
- *         &lt;element name="cputime" type="{http://www.w3.org/2001/XMLSchema}int" minOccurs="0"/>
- *         &lt;element name="stdout" type="{http://www.w3.org/2001/XMLSchema}string" minOccurs="0"/>
- *         &lt;element name="stderr" type="{http://www.w3.org/2001/XMLSchema}string" minOccurs="0"/>
- *         &lt;choice>
- *           &lt;element name="cmd" type="{http://www.w3.org/2001/XMLSchema}string"/>
- *           &lt;element name="container" type="{}tContainer"/>
- *         &lt;/choice>
- *       &lt;/sequence>
- *     &lt;/restriction>
- *   &lt;/complexContent>
- * &lt;/complexType>
- * </pre>
- * 
- * 
- */
-@XmlAccessorType(XmlAccessType.FIELD)
-@XmlType(name = "", propOrder = {
-    "user",
-    "cores",
-    "memory",
-    "cputime",
-    "stdout",
-    "stderr",
-    "cmd",
-    "container"
-})
-@XmlRootElement(name = "task")
-public class Task {
-
-    @XmlElement(required = true)
-    protected String user;
-    protected Integer cores;
-    protected Integer memory;
-    protected Integer cputime;
-    protected String stdout;
-    protected String stderr;
-    protected String cmd;
-    protected TContainer container;
-
-    /**
-     * Gets the value of the user property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link String }
-     *     
-     */
-    public String getUser() {
-        return user;
-    }
-
-    /**
-     * Sets the value of the user property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link String }
-     *     
-     */
-    public void setUser(String value) {
-        this.user = value;
-    }
-
-    /**
-     * Gets the value of the cores property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link Integer }
-     *     
-     */
-    public Integer getCores() {
-        return cores;
-    }
-
-    /**
-     * Sets the value of the cores property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link Integer }
-     *     
-     */
-    public void setCores(Integer value) {
-        this.cores = value;
-    }
-
-    /**
-     * Gets the value of the memory property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link Integer }
-     *     
-     */
-    public Integer getMemory() {
-        return memory;
-    }
-
-    /**
-     * Sets the value of the memory property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link Integer }
-     *     
-     */
-    public void setMemory(Integer value) {
-        this.memory = value;
-    }
-
-    /**
-     * Gets the value of the cputime property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link Integer }
-     *     
-     */
-    public Integer getCputime() {
-        return cputime;
-    }
-
-    /**
-     * Sets the value of the cputime property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link Integer }
-     *     
-     */
-    public void setCputime(Integer value) {
-        this.cputime = value;
-    }
-
-    /**
-     * Gets the value of the stdout property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link String }
-     *     
-     */
-    public String getStdout() {
-        return stdout;
-    }
-
-    /**
-     * Sets the value of the stdout property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link String }
-     *     
-     */
-    public void setStdout(String value) {
-        this.stdout = value;
-    }
-
-    /**
-     * Gets the value of the stderr property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link String }
-     *     
-     */
-    public String getStderr() {
-        return stderr;
-    }
-
-    /**
-     * Sets the value of the stderr property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link String }
-     *     
-     */
-    public void setStderr(String value) {
-        this.stderr = value;
-    }
-
-    /**
-     * Gets the value of the cmd property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link String }
-     *     
-     */
-    public String getCmd() {
-        return cmd;
-    }
-
-    /**
-     * Sets the value of the cmd property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link String }
-     *     
-     */
-    public void setCmd(String value) {
-        this.cmd = value;
-    }
-
-    /**
-     * Gets the value of the container property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link TContainer }
-     *     
-     */
-    public TContainer getContainer() {
-        return container;
-    }
-
-    /**
-     * Sets the value of the container property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link TContainer }
-     *     
-     */
-    public void setContainer(TContainer value) {
-        this.container = value;
-    }
-
-}
